// features/complain/complain-view/types.ts
export interface Complain {
    cust_first_name: string
    cust_last_name: string
    cust_id: string
    cust_address: string
    cust_landmark: string
    cust_phone_home: string
    cust_phone_office: string
    cust_phone_mobile: string
    email: string
    tech_id: string
    tech_name: string
    call_visit_tm_in:string,
    call_visit_tm_out:string,
    bill_no:string,
    grand_total:string,
    appoint_date:string,
    dnd:string,
    call_type:string,
    call_at:string,
    svc_name:string,
    happy_status:string,
    appoint_time:string,
    bill_date:string,
    tech_address: string
    tech_phone: string
    tech_mobile: string
    tech_email: string
    service_category_id: string
    service_category_name: string
    product_id: string
    product_name: string
    product_category_id: string
    product_category_name: string
    service_id: string
    service_type: string
    brand: string
    model: string
    call_serial_no: string
    call_purchase_dt: string
    call_dealer_name: any
    customer_remarks: string
    call_engineer_remark: string
    symptom_id: string
    symptom_code: string
    defect_id: string
    defect_code: string
    repair_id: string
    repair_code: string
    long: string
    lat: string
    call_uid: string
    call_dt: string
    call_tm: string
    call_status: string
    call_visit_dt: string
    call_reason_pending: string
    dispatch_remarks: any
    cust_warranty: string
    tech_warranty: any
    photo: any
    online_payment: any
    online_amount: any
    feedback: any
  }
  